+++ src/Plugins.jl
@@ -131,9 +131,6 @@ function install(path::String, dest::String; force = false)
 
   isdir(Genie.config.path_plugins) || mkpath(Genie.config.path_plugins)
 
-  @show pwd()
-  @show dest
-
   depth = 0
   for (root, dirs, files) in walkdir(path)
     dest_path = joinpath(abspath(dest), splitpath(root)[end-depth:end]...)